+++ java/com/jcraft/jsch/DHGEX256.java
@@ -1,6 +1,7 @@
 /* -*-mode:java; c-basic-offset:2; indent-tabs-mode:nil -*- */
 /*
 Copyright (c) 2002-2014 ymnk, JCraft,Inc. All rights reserved.
+Copyright (c) 2018, D. R. Commander. All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided that the following conditions are met:
@@ -77,7 +78,7 @@ public class DHGEX256 extends KeyExchange{
 
     try{
       Class c=Class.forName(session.getConfig("sha-256"));
-      sha=(HASH)(c.newInstance());
+      sha=(HASH)(c.getConstructor().newInstance());
       sha.init();
     }
     catch(Exception e){
@@ -89,7 +90,7 @@ public class DHGEX256 extends KeyExchange{
 
     try{
       Class c=Class.forName(session.getConfig("dh"));
-      dh=(com.jcraft.jsch.DH)(c.newInstance());
+      dh=(com.jcraft.jsch.DH)(c.getConstructor().newInstance());
       dh.init();
     }
     catch(Exception e){
@@ -259,7 +260,7 @@ System.err.println("0x"+Integer.toHexString(g[iii]&0xff)+",");
 	SignatureRSA sig=null;
 	try{
 	  Class c=Class.forName(session.getConfig("signature.rsa"));
-	  sig=(SignatureRSA)(c.newInstance());
+	  sig=(SignatureRSA)(c.getConstructor().newInstance());
 	  sig.init();
 	}
 	catch(Exception e){
@@ -305,7 +306,7 @@ System.err.println("0x"+Integer.toHexString(g[iii]&0xff)+",");
 	SignatureDSA sig=null;
 	try{
 	  Class c=Class.forName(session.getConfig("signature.dss"));
-	  sig=(SignatureDSA)(c.newInstance());
+	  sig=(SignatureDSA)(c.getConstructor().newInstance());
 	  sig.init();
 	}
 	catch(Exception e){