+++ java/com/jcraft/jsch/KeyPair.java
@@ -1,6 +1,7 @@
 /* -*-mode:java; c-basic-offset:2; indent-tabs-mode:nil -*- */
 /*
 Copyright (c) 2002-2014 ymnk, JCraft,Inc. All rights reserved.
+Copyright (c) 2018, D. R. Commander. All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided that the following conditions are met:
@@ -369,7 +370,7 @@ public abstract class KeyPair{
     if(random==null){
       try{
 	Class c=Class.forName(JSch.getConfig("random"));
-        random=(Random)(c.newInstance());
+        random=(Random)(c.getConstructor().newInstance());
       }
       catch(Exception e){ System.err.println("connect: random "+e); }
     }
@@ -379,7 +380,7 @@ public abstract class KeyPair{
   private HASH genHash(){
     try{
       Class c=Class.forName(JSch.getConfig("md5"));
-      hash=(HASH)(c.newInstance());
+      hash=(HASH)(c.getConstructor().newInstance());
       hash.init();
     }
     catch(Exception e){
@@ -390,7 +391,7 @@ public abstract class KeyPair{
     try{
       Class c;
       c=Class.forName(JSch.getConfig("3des-cbc"));
-      cipher=(Cipher)(c.newInstance());
+      cipher=(Cipher)(c.getConstructor().newInstance());
     }
     catch(Exception e){
     }
@@ -436,7 +437,7 @@ public abstract class KeyPair{
       }
       else if(vendor==VENDOR_PUTTY){
         Class c=Class.forName((String)JSch.getConfig("sha-1"));
-        HASH sha1=(HASH)(c.newInstance());
+        HASH sha1=(HASH)(c.getConstructor().newInstance());
         tmp = new byte[4];
         key = new byte[20*2];
         for(int i = 0; i < 2; i++){
@@ -648,7 +649,7 @@ public abstract class KeyPair{
           i+=8;
           if(Session.checkCipher((String)JSch.getConfig("aes256-cbc"))){
             Class c=Class.forName((String)JSch.getConfig("aes256-cbc"));
-            cipher=(Cipher)(c.newInstance());
+            cipher=(Cipher)(c.getConstructor().newInstance());
             // key=new byte[cipher.getBlockSize()];
             iv=new byte[cipher.getIVSize()];
           }
@@ -662,7 +663,7 @@ public abstract class KeyPair{
           i+=8;
           if(Session.checkCipher((String)JSch.getConfig("aes192-cbc"))){
             Class c=Class.forName((String)JSch.getConfig("aes192-cbc"));
-            cipher=(Cipher)(c.newInstance());
+            cipher=(Cipher)(c.getConstructor().newInstance());
             // key=new byte[cipher.getBlockSize()];
             iv=new byte[cipher.getIVSize()];
           }
@@ -676,7 +677,7 @@ public abstract class KeyPair{
           i+=8;
           if(Session.checkCipher((String)JSch.getConfig("aes128-cbc"))){
             Class c=Class.forName((String)JSch.getConfig("aes128-cbc"));
-            cipher=(Cipher)(c.newInstance());
+            cipher=(Cipher)(c.getConstructor().newInstance());
             // key=new byte[cipher.getBlockSize()];
             iv=new byte[cipher.getIVSize()];
           }
@@ -1026,7 +1027,7 @@ public abstract class KeyPair{
       if(Session.checkCipher((String)JSch.getConfig("aes256-cbc"))){
         try {
           Class c=Class.forName((String)JSch.getConfig("aes256-cbc"));
-          kpair.cipher=(Cipher)(c.newInstance());
+          kpair.cipher=(Cipher)(c.getConstructor().newInstance());
           kpair.iv=new byte[kpair.cipher.getIVSize()];
         }
         catch(Exception e){