+++ java/com/jcraft/jsch/KnownHosts.java
@@ -1,6 +1,7 @@
 /* -*-mode:java; c-basic-offset:2; indent-tabs-mode:nil -*- */
 /*
 Copyright (c) 2002-2014 ymnk, JCraft,Inc. All rights reserved.
+Copyright (c) 2018, D. R. Commander. All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided that the following conditions are met:
@@ -489,7 +490,7 @@ loop:
     if(hmacsha1==null){
       try{
         Class c=Class.forName(JSch.getConfig("hmac-sha1"));
-        hmacsha1=(MAC)(c.newInstance());
+        hmacsha1=(MAC)(c.getConstructor().newInstance());
       }
       catch(Exception e){ 
         System.err.println("hmacsha1: "+e);