+++ java/com/jcraft/jsch/Session.java
@@ -1,6 +1,7 @@
 /* -*-mode:java; c-basic-offset:2; indent-tabs-mode:nil -*- */
 /*
 Copyright (c) 2002-2014 ymnk, JCraft,Inc. All rights reserved.
+Copyright (c) 2018, D. R. Commander. All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided that the following conditions are met:
@@ -192,7 +193,7 @@ public class Session implements Runnable{
     if(random==null){
       try{
 	Class c=Class.forName(getConfig("random"));
-        random=(Random)(c.newInstance());
+        random=(Random)(c.getConstructor().newInstance());
       }
       catch(Exception e){ 
         throw new JSchException(e.toString(), e);
@@ -380,7 +381,7 @@ public class Session implements Runnable{
       UserAuth ua=null;
       try{
 	Class c=Class.forName(getConfig("userauth.none"));
-        ua=(UserAuth)(c.newInstance());
+        ua=(UserAuth)(c.getConstructor().newInstance());
       }
       catch(Exception e){ 
         throw new JSchException(e.toString(), e);
@@ -447,7 +448,7 @@ public class Session implements Runnable{
             Class c=null;
             if(getConfig("userauth."+method)!=null){
               c=Class.forName(getConfig("userauth."+method));
-              ua=(UserAuth)(c.newInstance());
+              ua=(UserAuth)(c.getConstructor().newInstance());
             }
           }
           catch(Exception e){
@@ -592,7 +593,7 @@ public class Session implements Runnable{
     KeyExchange kex=null;
     try{
       Class c=Class.forName(getConfig(guess[KeyExchange.PROPOSAL_KEX_ALGS]));
-      kex=(KeyExchange)(c.newInstance());
+      kex=(KeyExchange)(c.getConstructor().newInstance());
     }
     catch(Exception e){ 
       throw new JSchException(e.toString(), e);
@@ -1136,7 +1137,7 @@ key_type+" key fingerprint is "+key_fprint+".\n"+
   
       method=guess[KeyExchange.PROPOSAL_ENC_ALGS_STOC];
       c=Class.forName(getConfig(method));
-      s2ccipher=(Cipher)(c.newInstance());
+      s2ccipher=(Cipher)(c.getConstructor().newInstance());
       while(s2ccipher.getBlockSize()>Es2c.length){
         buf.reset();
         buf.putMPInt(K);
@@ -1154,7 +1155,7 @@ key_type+" key fingerprint is "+key_fprint+".\n"+
 
       method=guess[KeyExchange.PROPOSAL_MAC_ALGS_STOC];
       c=Class.forName(getConfig(method));
-      s2cmac=(MAC)(c.newInstance());
+      s2cmac=(MAC)(c.getConstructor().newInstance());
       MACs2c = expandKey(buf, K, H, MACs2c, hash, s2cmac.getBlockSize());
       s2cmac.init(MACs2c);
       //mac_buf=new byte[s2cmac.getBlockSize()];
@@ -1163,7 +1164,7 @@ key_type+" key fingerprint is "+key_fprint+".\n"+
 
       method=guess[KeyExchange.PROPOSAL_ENC_ALGS_CTOS];
       c=Class.forName(getConfig(method));
-      c2scipher=(Cipher)(c.newInstance());
+      c2scipher=(Cipher)(c.getConstructor().newInstance());
       while(c2scipher.getBlockSize()>Ec2s.length){
         buf.reset();
         buf.putMPInt(K);
@@ -1181,7 +1182,7 @@ key_type+" key fingerprint is "+key_fprint+".\n"+
 
       method=guess[KeyExchange.PROPOSAL_MAC_ALGS_CTOS];
       c=Class.forName(getConfig(method));
-      c2smac=(MAC)(c.newInstance());
+      c2smac=(MAC)(c.getConstructor().newInstance());
       MACc2s = expandKey(buf, K, H, MACc2s, hash, c2smac.getBlockSize());
       c2smac.init(MACc2s);
 
@@ -2207,7 +2208,7 @@ break;
          (isAuthed && method.equals("zlib@openssh.com"))){
         try{
           Class c=Class.forName(foo);
-          deflater=(Compression)(c.newInstance());
+          deflater=(Compression)(c.getConstructor().newInstance());
           int level=6;
           try{ level=Integer.parseInt(getConfig("compression_level"));}
           catch(Exception ee){ }
@@ -2234,7 +2235,7 @@ break;
          (isAuthed && method.equals("zlib@openssh.com"))){
         try{
           Class c=Class.forName(foo);
-          inflater=(Compression)(c.newInstance());
+          inflater=(Compression)(c.getConstructor().newInstance());
           inflater.init(Compression.INFLATER, 0);
         }
         catch(Exception ee){
@@ -2469,7 +2470,7 @@ break;
   static boolean checkCipher(String cipher){
     try{
       Class c=Class.forName(cipher);
-      Cipher _c=(Cipher)(c.newInstance());
+      Cipher _c=(Cipher)(c.getConstructor().newInstance());
       _c.init(Cipher.ENCRYPT_MODE,
               new byte[_c.getBlockSize()],
               new byte[_c.getIVSize()]);
@@ -2514,7 +2515,7 @@ break;
   static boolean checkKex(Session s, String kex){
     try{
       Class c=Class.forName(kex);
-      KeyExchange _c=(KeyExchange)(c.newInstance());
+      KeyExchange _c=(KeyExchange)(c.getConstructor().newInstance());
       _c.init(s ,null, null, null, null);
       return true;
     }